refactor: implement password reset at the service layer
